#50 Make TextGraphWriter an inner class of TextGraphFormatter
